+++ pybuoy/const.py
@@ -7,6 +7,7 @@ from pybuoy.endpoints import API_PATH as API_PATH  # noqa: F401
 
 class Endpoints(Enum):
     ACTIVE_STATIONS = "active_stations"
+    FORECASTS = "forecasts"
     REALTIME = "realtime2"