+++ pybuoy/mixins/parser.py
@@ -1,11 +1,14 @@
 """Provide the ParserMixin class."""
 from collections import defaultdict
 from datetime import datetime as dt
+from os.path import dirname, join
 from typing import Any
 from xml.etree.ElementTree import Element, fromstring
 
+import lxml.etree as ET  # type: ignore
+
 from pybuoy.const import RealtimeDatasets, RealtimeDatasetsValues
-from pybuoy.exceptions import BuoyException
+from pybuoy.exceptions import NDBCException
 from pybuoy.observation import MeteorologicalObservation, WaveSummaryObservation
 from pybuoy.observation.observations import (
     MeteorologicalObservations,
@@ -17,9 +20,12 @@ from pybuoy.unit_mappings import MeteorologicalKey, WaveSummaryKey
 class ParserMixin:
     """Parser mixin supports handling of third-party data."""
 
+    def __init__(self) -> None:
+        self.__set_xslt_transform()
+
     def etree_to_dict(self, element: Element):
         """Parse XML Element to dictionary."""
-        from_etree: dict[str, dict[str, Any] | Any] = {  # TODO: improve typing
+        from_etree: dict[str, dict[str, Any] | Any] = {
             element.tag: {} if element.attrib else None
         }
         children = list(element)
@@ -30,11 +36,14 @@ class ParserMixin:
                     parsed_children[key].append(value)
             from_etree = {
                 element.tag: {
-                    k: v[0] if len(v) == 1 else v for k, v in parsed_children.items()
+                    key: value[0] if len(value) == 1 else value
+                    for key, value in parsed_children.items()
                 }
             }
         if element.attrib:
-            from_etree[element.tag].update((k, v) for k, v in element.attrib.items())
+            from_etree[element.tag].update(
+                (key, value) for key, value in element.attrib.items()
+            )
         if element.text:
             text = element.text.strip()
             if children or element.attrib:
@@ -52,7 +61,7 @@ class ParserMixin:
         obs = self.__clean_realtime_data(data=data, dataset=dataset)
         # TODO: refine error handling
         if obs is None or len(obs) == 0:
-            raise BuoyException
+            raise NDBCException
 
         match dataset:
             case RealtimeDatasets.spec.value:
@@ -65,18 +74,23 @@ class ParserMixin:
             case _:
                 return data
 
+    @property
+    def xslt_transform(self) -> ET.XSLT:
+        return self._xslt_transform
+
     # TODO: consider station class/dto like observations
     def _clean_activestation_data(self, data: str):
         xml_tree_root = fromstring(data)
         # TODO: consider incorporating `etree_to_dict`
         return [dict(el.items()) for el in xml_tree_root.findall("station")]
 
+    # TODO: fix missing latest row
     def __clean_realtime_data(
         self, data: str, dataset: RealtimeDatasetsValues
     ) -> list[MeteorologicalObservation] | list[WaveSummaryObservation]:
         if data is None:
             # TODO: handle when request not successful
-            raise BuoyException
+            raise NDBCException
 
         # TODO: consider csv module
         rows = data.strip().split("\n")
@@ -94,6 +108,7 @@ class ParserMixin:
                 int(record_array[3]),
                 int(record_array[4]),
             )
+            # TODO: investigate adding Strategy Pattern here
             if dataset == RealtimeDatasets.txt.value:
                 observation = self.__parse_meteorological_record(
                     date_recorded=date_recorded,
@@ -110,16 +125,19 @@ class ParserMixin:
 
         return obs
 
-    # TODO: refactor
     def __parse_meteorological_record(self, date_recorded: dt, headers, records):
         values: dict[MeteorologicalKey, str] = {
             MeteorologicalKey[headers[idx]]: value for idx, value in enumerate(records)
         }
         return MeteorologicalObservation(values, date_recorded)
 
-    # TODO: combine with above
+    # TODO: refactor (e.g., combine with __parse_meteorological_record)
     def __parse_wave_summary_record(self, date_recorded: dt, headers, records):
         values: dict[WaveSummaryKey, str] = {
             WaveSummaryKey[headers[idx]]: value for idx, value in enumerate(records)
         }
         return WaveSummaryObservation(values, date_recorded)
+
+    def __set_xslt_transform(self, xsl_filename="noaa.xsl"):
+        forecast_xslt = ET.parse(join(dirname(__file__), xsl_filename))
+        self._xslt_transform = ET.XSLT(forecast_xslt)