+++ pybuoy/observation/__init__.py
@@ -1,10 +1,14 @@
 from pybuoy.observation.observation import MeteorologicalObservation  # noqa: F401
-from pybuoy.observation.observation import WaveSummaryObservation  # noqa: F401
+from pybuoy.observation.observation import (  # noqa: F401
+    ForecastObservation,
+    WaveSummaryObservation,
+)
 from pybuoy.observation.observation_datum import (  # noqa: F401
     ObservationFloatDatum,
     ObservationStringDatum,
 )
 from pybuoy.observation.observations import (  # noqa: F401
+    ForecastObservations,
     MeteorologicalObservations,
     WaveSummaryObservations,
 )